Refactor : #3 - JWT 토큰 관련 Context API 모듈화
로그인 유무를 판별하는 Context API 모듈을 별도의 폴더로 저장하였습니다.
