+++ src/components/common/Navbar.jsx
@@ -2,13 +2,13 @@ import styled, { css } from "styled-components"
 import { useContext } from "react"
 import { Link } from "react-router-dom"
 
-import { AuthContext } from "pages/MainPage"
+import { AuthStateContext } from "module/Auth"
 
-const Navbar = ({ textColor }) => {
-  const { authState } = useContext(AuthContext)
+const Navbar = () => {
+  const authState = useContext(AuthStateContext)
   const { authenticated } = authState
   return (
-    <NavbarLayout textColor={textColor}>
+    <NavbarLayout>
       <LinkElement to="/">Main</LinkElement>
       {authenticated || (
         <>
@@ -21,7 +21,7 @@ const Navbar = ({ textColor }) => {
 }
 
 const NavbarLayout = styled.ul`
-  ${({ textColor }) => {
+  ${({ theme }) => {
     return css`
       width: 40vw;
       height: 3vw;
@@ -33,7 +33,7 @@ const NavbarLayout = styled.ul`
       justify-content: space-between;
 
       text-align: center;
-      color: ${textColor};
+      color: ${theme.colors.white};
     `
   }}
 `
