+++ src/index.js
@@ -2,6 +2,7 @@ import React from "react"
 import ReactDOM from "react-dom/client"
 import App from "./App"
 
+import { AuthProvider } from "module/Auth"
 import { ThemeProvider } from "styled-components"
 import { defalutTheme as theme } from "styles/Theme"
 import GlobalStyle from "styles/GlobalStyle"
@@ -9,7 +10,9 @@ import GlobalStyle from "styles/GlobalStyle"
 const root = ReactDOM.createRoot(document.getElementById("root"))
 root.render(
   <ThemeProvider theme={theme}>
-    <GlobalStyle />
-    <App />
+    <AuthProvider>
+      <GlobalStyle />
+      <App />
+    </AuthProvider>
   </ThemeProvider>,
 )
