+++ src/pages/MainPage.jsx
@@ -1,24 +1,11 @@
-import { createContext, useEffect, useReducer } from "react"
+import { useContext, useEffect } from "react"
 import { BrowserRouter, Routes, Route, Outlet, Navigate } from "react-router-dom"
 
 import BaseTemplate from "components/template/BaseTemplate"
 import { HomeContainer } from "pages/Container/HomeContainer"
 import { LoginContainer } from "pages/Container/LoginContainer"
 import { SignupContainer } from "pages/Container/SignupContainer"
-
-// JWT Token 관련 reducer 함수, initialState 객체 선언.
-const initialState = { token: null, authenticated: false }
-const reducer = (state, action) => {
-  switch (action.type) {
-    case "SET_TOKEN":
-      return { ...state, token: action.token, authenticated: true }
-    case "REMOVE_TOKEN":
-      return { ...state, token: null, authenticated: false }
-    default:
-      return state
-  }
-}
-export const AuthContext = createContext()
+import { AuthDispatchContext, AuthStateContext } from "module/Auth"
 
 // 오직 로그인이 되었을때만 접근이 가능하도록 하는 Route
 const PrivateRoute = ({ isLogin }) => {
@@ -31,11 +18,10 @@ const RestrictedRoute = ({ isLogin }) => {
 }
 
 const MainPage = () => {
-  const [authState, authDispatch] = useReducer(reducer, initialState)
-  const { authenticated } = authState
-
-  /* 화면이 새로 로딩될 때, localStorage에 access_token이 있는지를 체크.
-     만약 토큰이 있다면 이를 자동으로 Context API에 추가하도록 설정. */
+  const authState = useContext(AuthStateContext)
+  const authDispatch = useContext(AuthDispatchContext)
+  //  화면이 새로 로딩될 때, localStorage에 access_token이 있는지를 체크.
+  //  만약 토큰이 있다면 이를 자동으로 Context API에 추가하도록 설정.
   useEffect(() => {
     const accessToken = localStorage.getItem("access_token")
     if (accessToken) {
@@ -47,19 +33,17 @@ const MainPage = () => {
   }, [])
 
   return (
-    <AuthContext.Provider value={{ authState, authDispatch }}>
-      <BrowserRouter>
-        <BaseTemplate>
-          <Routes>
-            <Route path="/" element={<HomeContainer />} />
-            <Route element={<RestrictedRoute />} isLogin={authenticated}>
-              <Route path="/login" element={<LoginContainer />} />
-              <Route path="/signup" element={<SignupContainer />} />
-            </Route>
-          </Routes>
-        </BaseTemplate>
-      </BrowserRouter>
-    </AuthContext.Provider>
+    <BrowserRouter>
+      <BaseTemplate>
+        <Routes>
+          <Route path="/" element={<HomeContainer />} />
+          <Route element={<RestrictedRoute />} isLogin={authState.authenticated}>
+            <Route path="/login" element={<LoginContainer />} />
+            <Route path="/signup" element={<SignupContainer />} />
+          </Route>
+        </Routes>
+      </BaseTemplate>
+    </BrowserRouter>
   )
 }
